added browse_foreigners.lib.php and extracted methods from browse_foreigners to it.
